requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.21.0
python-dotenv>=0.19.0

# Forensic and extraction
python-magic>=0.4.27
PyYAML>=6.0

# OCR and image processing
pytesseract>=0.3.10
Pillow>=9.0.0

# Sentiment analysis
textblob>=0.17.1

# Reporting
python-docx>=0.8.11
reportlab>=3.6.0
openpyxl>=3.0.9
XlsxWriter>=3.0.2

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Optional for enhanced features
azure-ai-textanalytics>=5.2.0
